Add initial app structure with placeholder functionality
This commit introduces the basic structure of a React application with TypeScript. It contains mockups of components for different homework tasks. Dummy functionality is indicated with comments stipulating what needs to be done. Each homework task (HW) has its own directory and associated files. Also added are a .gitignore file to exclude unwanted files and directories as well as configuration files for TypeScript and Prettier.