Enable HW1 functionality and update package.json
This commit handles two key actions. Firstly, it updates the 'src/App.tsx' and 'src/HW/1/HW1.tsx' files to enable the functionality of 'HW1' which includes the display of a user list. Secondly, 'package.json' has been updated to include 'homepage' and 'gh-pages' dependencies for deployment purposes, along with the addition of 'predeploy' and 'deploy' scripts. It also updates the 'yarn.lock' file due to the new dependencies in package.json.